+++ Main.Semantics.cs
@@ -39,6 +39,7 @@ namespace DDF___Program_Language_Editor
 
             leaveFoldedView();
             hideCompletion();
+            hideSignatureHelp();
             string newName = requestSymbolRename(occurrence.Symbol.Name);
             if (newName == null) return;